+++ whittle/lora_model/merge.py
@@ -1,0 +1,25 @@
+from pathlib import Path
+def merge_lora(
+    sampling_strategy: str,
+    checkpoint_dir: Path,
+    pretrained_checkpoint_dir: str | None = None,
+    precision: str | None = None,
+) -> None:
+    """Merges the LoRA weights with the base model.
+
+    See ``litgpt finetune lora``.
+
+    Creates a new ``lit_model.pth`` file by merging the LoRA weights (``lit_model.pth.lora``)
+    with the original checkpoint weights.
+
+    Arguments:
+        checkpoint_dir: Path to the checkpoint directory with trained LoRA weights, which is the output of
+            ``litgpt finetune lora``.
+        pretrained_checkpoint_dir: Optional path to the checkpoint directory with the weights of the base model
+            corresponding to the LoRA checkpoint. By default, this will automatically be inferred from the metadata
+            in the given `checkpoint_dir` directory. Only set this if the base model's checkpoint directory
+            has moved or was renamed.
+        precision: Optional precision setting to instantiate the model weights in. By default, this will
+            automatically be inferred from the metadata in the given ``checkpoint_dir`` directory.
+    """
+    raise NotImplementedError("Merging is currently not supported.")
